Move the helper function back in experiment_server.py
