+++ dallinger/experiment_server/experiment_server.py
@@ -14,7 +14,6 @@ from flask import (
     request,
     Response,
     send_from_directory,
-    url_for,
 )
 from flask_login import current_user, LoginManager, login_required
 from jinja2 import TemplateNotFound
@@ -27,11 +26,12 @@ from psycopg2.extensions import TransactionRollbackError
 
 from dallinger import db
 from dallinger import experiment
-from dallinger import models
-from dallinger.config import get_config
-from dallinger import recruiters
 from dallinger.notifications import admin_notifier
 from dallinger.notifications import MessengerError
+from dallinger.config import get_config
+from dallinger.utils import generate_random_id
+from dallinger import models
+from dallinger import recruiters
 
 from . import dashboard
 from .replay import ReplayBackend
@@ -43,8 +43,12 @@ from .utils import (
     error_response,
     success_response,
     ExperimentError,
-    prepare_advertisement
+    ValidatesBrowser
 )
+
+from flask import redirect
+from flask import url_for
+from json import dumps
 
 
 # Initialize the Dallinger database.
@@ -461,6 +465,65 @@ def launch():
     )
 
     return success_response(recruitment_msg=message)
+
+def prepare_advertisement():
+    session = db.session
+    config = _config()
+    mode = config.get("mode")
+
+    # Browser rule validation, if configured:
+    browser = ValidatesBrowser(config)
+    if not browser.is_supported(request.user_agent.string):
+        raise ExperimentError("browser_type_not_allowed")
+
+    entry_information = request.args.to_dict()
+
+    if entry_information.get("generate_tokens", None) in ("1", "true", "yes"):
+        redirect_params = entry_information.copy()
+        del redirect_params["generate_tokens"]
+        for entry_param in ("hitId", "assignmentId", "workerId"):
+            if not redirect_params.get(entry_param):
+                redirect_params[entry_param] = generate_random_id()
+        return True, {'redirect': redirect(url_for("advertisement", **redirect_params))}
+
+    app_id = config.get("id", "unknown")
+    exp = Experiment(session)
+    entry_data = exp.normalize_entry_information(entry_information)
+
+    hit_id = entry_data.get("hit_id")
+    assignment_id = entry_data.get("assignment_id")
+    worker_id = entry_data.get("worker_id")
+
+    if not (hit_id and assignment_id):
+        raise ExperimentError("hit_assign_worker_id_not_set_by_recruiter")
+
+    if worker_id is not None:
+        # Check if this workerId has completed the task before
+        already_participated = (
+                models.Participant.query.filter(
+                    models.Participant.worker_id == worker_id
+                ).first()
+                is not None
+        )
+
+        if already_participated:
+            raise ExperimentError("already_did_exp_hit")
+
+    recruiter_name = request.args.get("recruiter")
+    if not recruiter_name:
+        recruiter = recruiters.from_config(config)
+        recruiter_name = recruiter.nickname
+
+    kwargs = {
+        'recruiter': recruiter_name,
+        'hitid': hit_id,
+        'assignmentid': assignment_id,
+        'workerid': worker_id,
+        'mode': mode,
+        'app_id': app_id,
+        'query_string': request.query_string.decode()
+    }
+    return False, kwargs
 
 
 @app.route("/ad", methods=["GET"])
